dtype prob?

+++ bingo_table_maker_nth_inanna_neverbingo.py
@@ -240,7 +240,7 @@ def fill_table(initial_tables, nth_inanna, tid, use_tqdm=False):
                     else:
                         if step<3*WEAK_LIMIT:
                             next_loop = (loop_id 
-                                        -((25**(3*WEAK_LIMIT-1-step))
+                                        -(int(25**(3*WEAK_LIMIT-1-step))
                                             *(5*action_x+action_y))
                                         -1 #self 1
                                     )
